producer/requirements.txt
pandas==2.2.2          
kafka-python==2.0.2    
tqdm==4.66.3           